BREAKING: Move ImageRef & PlatformTagSet to common.docker module to avoid circular imports
* exceptions & types must be at the lowest layer of import tree.
* logging & validators come next and other per-situ utilities come last.
